P14 (part 4): the floor an owner recognises as their own room
The screens, and the session's last piece. A Floor rail item, because
it answers a different question from the billing grid: not "which
table am I putting this dosa on" but "which table needs me" — audit F5
is that second question going unanswered.

* the **plan**, dragged onto a grid of squares, with the section grid
  as a fallback that is not a degraded mode: no shop should have to
  draw a floor plan before it can bill;
* **two timers on every tile** — how long the table has been sitting,
  and how long since food went to the kitchen — with the thresholds
  read from the shop's own settings and the tile state arriving
  DECIDED, so no screen compares a number to a threshold;
* the **occupancy line** as sentences Rust wrote;
* filters by section and by "needs attention";
* **the room setup**: sections, a numbered run of tables, the table
  list showing what each one PRINTS as, hide/delete with the reason
  attached, and the two thresholds;
* **move and merge** from the tile.

A fifth tile state came with it (`waiting`), because one threshold
meant a table was fine and then suddenly red — a smoke alarm rather
than a floor view. It differs in form as well as colour: stripe only,
no wash. The kit test asserting four states asserts five now.

RUN IT AND LOOK AT IT, and this time the whole loop:
  * a dosa on table 11, moved to table 1 — same bill number, same
    food, both timers ticking, table 11 free;
  * an idli on table 2, merged into table 1: 126.00 + 84.00 = 210.00
    on one tile and table 2 free;
  * `peek` read the disk back and said what the screen said — the
    absorbed order is CANCELLED, keeps bill 0002, and carries
    merged_into pointing at the survivor. order_events has both
    kitchen tickets, the move and the merge, with who did them;
  * setting the late threshold to 2 minutes turned the tile red on the
    floor AND on the billing grid, which is what sharing it is for.

One bug found by looking, and it is audit F8 again: a toast said "The
order moved to tbl_1" — a shopkeeper reading our key. It says the
table's printed name now.

`peek` learned to show merges, splits, order_events and the floor
plan; it could not show what this session built, which is the third
time it has needed extending for exactly that reason.

694 tests (575 Rust + 119 front-end), clippy clean, all three
front-end guards clean.

+++ src-tauri/src/floor_tests.rs
@@ -420,8 +420,8 @@ fn the_two_timers_come_from_settings() {
     // The order was created an hour ago in fixture time, which is the past
     // relative to a real clock, so it is comfortably "late" at any threshold.
     let before = floor_on(&app).expect("the floor");
-    assert_eq!(before.warn_minutes, crate::floor::DEFAULT_WARN_MINUTES);
-    assert_eq!(before.late_minutes, crate::floor::DEFAULT_LATE_MINUTES);
+    assert_eq!(i64::from(before.warn_minutes), crate::floor::DEFAULT_WARN_MINUTES);
+    assert_eq!(i64::from(before.late_minutes), crate::floor::DEFAULT_LATE_MINUTES);
 
     let after = save_thresholds_on(&app, 5, 10).expect("saved");
     assert_eq!(after.warn_minutes, 5);
